requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pdfplumber==0.11.5
python-docx==1.1.2
sqlalchemy==2.0.38
httpx==0.28.1
pydantic==2.10.6

# Testing
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.0.0
respx==0.22.0
